requirements.txt
Pillow
opencv-python
numpy
pyocr
aiohttp
asyncio
discord.py
